+++ celestra-math.js
@@ -1,9 +1,9 @@
 /**
  * @name Celestra Math plugin
- * @version 5.3.0
+ * @version 5.3.1
  * @see https://github.com/Serrin/Celestra
  * @license MIT https://opensource.org/licenses/MIT
- * Required Celestra version: 5.3.0
+ * Required Celestra version: 5.3.1
  */
 
 (function(celestra){